Refactor: pull out triggers
