+++ .cjsescache
@@ -4,5 +4,10 @@
   "node_modules/webextension-polyfill/dist/browser-polyfill.js",
   "src/lib/extension-pref.js",
   "src/lib/pref-body.js",
-  "src/lib/pref-default.js"
+  "src/lib/pref-default.js",
+  "src/lib/triggers/click.mjs",
+  "src/lib/triggers/hover.mjs",
+  "src/lib/triggers/index.mjs",
+  "src/lib/triggers/load.mjs",
+  "src/lib/triggers/mutation.mjs"
 ]
