+++ src/lib/triggers/cache.mjs
@@ -1,0 +1,3 @@
+export const processedNodes = new WeakSet;
+export const nodeValidationCache = new WeakMap; // Node -> boolean
+
