+++ src/lib/triggers/index.mjs
@@ -1,0 +1,6 @@
+import load from "./load.mjs"
+import click from "./click.mjs"
+import hover from "./hover.mjs"
+import mutation from "./mutation.mjs";
+
+export default [load, click, hover, mutation];
